Add custom user ignores and split out checks.
